+++ app/src/main/java/com/and04/naturealbum/ui/album/AlbumViewModel.kt
@@ -3,7 +3,7 @@ package com.and04.naturealbum.ui.album
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.and04.naturealbum.data.dto.AlbumDto
-import com.and04.naturealbum.data.repository.DataRepository
+import com.and04.naturealbum.data.repository.local.LocalDataRepository
 import com.and04.naturealbum.ui.model.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -13,7 +13,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class AlbumViewModel @Inject constructor(
-    private val repository: DataRepository,
+    private val repository: LocalDataRepository,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow<UiState<List<AlbumDto>>>(UiState.Idle)
     val uiState: StateFlow<UiState<List<AlbumDto>>> = _uiState
