+++ app/src/main/java/com/and04/naturealbum/ui/photoinfo/PhotoInfoViewModel.kt
@@ -2,9 +2,8 @@ package com.and04.naturealbum.ui.photoinfo
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.and04.naturealbum.data.repository.DataRepository
 import com.and04.naturealbum.data.repository.RetrofitRepository
-import com.and04.naturealbum.data.room.Label
+import com.and04.naturealbum.data.repository.local.LocalDataRepository
 import com.and04.naturealbum.data.room.PhotoDetail
 import com.and04.naturealbum.ui.model.AlbumData
 import com.and04.naturealbum.ui.model.UiState
@@ -16,7 +15,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class PhotoInfoViewModel @Inject constructor(
-    private val roomRepository: DataRepository,
+    private val roomRepository: LocalDataRepository,
     private val retrofitRepository: RetrofitRepository,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow<UiState<AlbumData>>(UiState.Idle)
@@ -38,7 +37,7 @@ class PhotoInfoViewModel @Inject constructor(
         }
     }
 
-    fun setAlbumThumbnail(photoDetailId: Int){
+    fun setAlbumThumbnail(photoDetailId: Int) {
         viewModelScope.launch {
             roomRepository.updateAlbumPhotoDetailByAlbumId(photoDetailId)
         }
